added gtrig to the test-suite
For this, I improved the ergonomics of testsuite/src/main.rs:
the second parameter (manifest URL) can now be the name of a format,
in which case the correct manifest URL is automatically set.

This allows to use the same manifest URL for different formats
(in this case, TriG et GTriG).