add more RDF* support in the API
in rio_common
- rename NamedOrBlankNode to GraphName
- add type Subject, which is NamedOrBlankNode + Triple (if 'star' enabled)
- Subject and Term are marked as non_exhaustive enums;
  this is a breaking change, but allows non-star crates (e.g. rio_xml)
  to co-exist with star crates

in other crates:
- propagate the changes above
- in particular, because Subject and Term are now non-exhaustive enums,
  some From implementations have been replaced with TryFrom;
  this should be zero-cost when all existing variants are actually
  handled.

NB: a consequence of adding the Triple variant is that
owned versions of Terms can no longer be converted into the borrowed versions
(or at least not easily).
It appears that such conversions were not really needed
(for example, they were used to compare owned terms with borrowed terms,
which is now achieved by implementing PartialEq instead).