Adds nested entity resolution to RdfXmlParser.
The file added to the testsuite shows an instance where the parser currently fails.
This commit changes `parser::parse_doctype()` so that the value of every entity is escaped before it is added to the set of custom entities.
A possible issue is that an entity containing a nested entity which is defined afterwards would still cause a failure.